use react-swipeable for better ux
react-swipeable uses the touch being removed for the swipe to activate
whereas react-touch was based on the distance from the initial point.

I think this logic is flawed, as it caused serious usability issues in
which, seemingly, the swipe would be interpreted as going in random
directions.

And, with that, react-touch is now removed